#backend: move cloning topics into topicrepository
